+++ controls/BCGControlBar/BCGToolBarImages.cpp
@@ -622,7 +622,7 @@ PBITMAPINFO CreateBitmapInfoStruct(HWND hwnd, HBITMAP hBmp)
      if (cClrBits != 24) 
          pbmi = (PBITMAPINFO) LocalAlloc(LPTR, 
                     sizeof(BITMAPINFOHEADER) + 
-                    sizeof(RGBQUAD) * (UINT)(1<< cClrBits)); 
+                    sizeof(RGBQUAD) * (1<< cClrBits)); 
 
      // There is no RGBQUAD array for the 24-bit-per-pixel format. 
 
@@ -1164,7 +1164,7 @@ static BOOL WriteDIB( LPCTSTR szFile, HANDLE hDIB)
 
 	// Fill in the fields of the file header 
 	hdr.bfType		= ((WORD) ('M' << 8) | 'B');	// is always "BM"
-	hdr.bfSize		= (DWORD)GlobalSize (hDIB) + sizeof( hdr );
+	hdr.bfSize		= GlobalSize (hDIB) + sizeof( hdr );
 	hdr.bfReserved1 	= 0;
 	hdr.bfReserved2 	= 0;
 	hdr.bfOffBits		= (DWORD) (sizeof( hdr ) + lpbi->biSize +
@@ -1174,7 +1174,7 @@ static BOOL WriteDIB( LPCTSTR szFile, HANDLE hDIB)
 	file.Write( &hdr, sizeof(hdr) );
 
 	// Write the DIB header and the bits 
-	file.Write( lpbi, (UINT)GlobalSize(hDIB) );
+	file.Write( lpbi, GlobalSize(hDIB) );
 
 	return TRUE;
 }
