+++ controls/VTDesign/3DBar.cpp
@@ -892,7 +892,7 @@ LRESULT CVT3DBar::DefWindowProc(UINT message, WPARAM wParam, LPARAM lParam)
 
 	if( message == WM_GETTEXT )
 	{
-		lstrcpyn((LPTSTR)lParam, (LPCTSTR)m_strText, (int)wParam);
+		lstrcpyn((LPTSTR)lParam, (LPCTSTR)m_strText, wParam);
 		if ((int)wParam > m_strText.GetLength())
 		wParam = m_strText.GetLength();
 		return TRUE;
